done for venues, fields, slots and message features

+++ src/main/java/com/badminton/shop/ws_booking_sport/goong/GoongResponse.java
@@ -25,7 +25,8 @@ public class GoongResponse {
     @Data
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class Location {
-        private float lat;
-        private float lng;
+        // Use double for higher precision
+        private double lat;
+        private double lng;
     }
 }
